Initial work on rewriting IMAP support in ingo.
This is 100% guaranteed not to work yet (I don't have setup to test at
the moment), although it is a start.  The 'mock' driver is guaranteed
never to work since I have no idea what the hell it does and there is
absolutely zero documentation on it.  Unless someone comes forward to
fix it, it will get the ax here very soon.